managed-connectivity/aws-glue-connector/requirements.txt
boto3
google-cloud-dataplex==2.4.0
google-cloud-storage
google-cloud-secret-manager